Rename serde to serdes
