+++ src/main/java/org/akhq/repositories/SchemaRegistryRepository.java
@@ -5,7 +5,7 @@ import io.confluent.kafka.schemaregistry.client.rest.entities.requests.ConfigUpd
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
 import io.confluent.kafka.serializers.KafkaAvroDeserializer;
 import org.akhq.models.Schema;
-import org.akhq.utils.avroserde.AvroSerializer;
+import org.akhq.utils.avroserdes.AvroSerializer;
 import org.akhq.modules.KafkaModule;
 import org.akhq.utils.PagedList;
 import org.akhq.utils.Pagination;
