+++ src/main/java/org/akhq/utils/avroserdes/AvroToJsonSerializer.java
@@ -1,17 +1,11 @@
-package org.akhq.utils.avroserde;
+package org.akhq.utils.avroserdes;
 
-import org.apache.avro.*;
-import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
-import org.apache.avro.io.Encoder;
 import org.apache.avro.io.EncoderFactory;
 import org.apache.avro.io.JsonEncoder;
-import org.apache.avro.specific.SpecificDatumWriter;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.math.BigDecimal;
-import java.nio.ByteBuffer;
 
 public class AvroToJsonSerializer {
